requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
requests>=2.31.0
